# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ozel Ders Evim API
  version: 0.1.0
  description: MVP API contract. Feature branches must update this file with implemented endpoints.
servers:
  - url: /api
paths:
...
  /auth/forgot-password:
    post:
      summary: Send a password reset email without revealing account existence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
...
  /teacher-eligibility/submissions:
    post:
      summary: Submit teacher eligibility test answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherEligibilitySubmissionPayload"
      responses:
        "200":
          description: Eligibility attempt scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherEligibilitySubmissionResponse"
        "400":
          description: Invalid, duplicate, or incomplete answers
        "401":
          description: Teacher is not signed in
        "403":
          description: Current account is not a teacher
        "404":
          description: Test not found
        "409":
          description: Teacher already passed the test
...
  /favorites/toggle:
    post:
      summary: Toggle favorite state for a teacher listing (student only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:
                  type: string
      responses:
        "200":
          description: New favorite state after toggle
          content:
            application/json:
              schema:
                type: object
                required: [isFavorited]
                properties:
                  isFavorited:
                    type: boolean
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a student account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Teacher listing not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
...
components:
  responses:
    RateLimitExceeded:
      description: Too many requests in the current time window
      headers:
        Retry-After:
          schema:
            type: string
        X-RateLimit-Remaining:
          schema:
            type: string
        X-RateLimit-Reset:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    TeacherEligibilitySubmissionPayload:
      type: object
      required: [testId, answers]
      properties:
        testId:
          type: string
          format: uuid
        answers:
          type: array
          items:
            type: object
            required: [questionId, choiceId]
            properties:
              questionId:
                type: string
              choiceId:
                type: string
    TeacherEligibilitySubmissionResponse:
      type: object
      required: [attemptId, status, score, canPublishListing]
      properties:
        attemptId:
          type: string
          format: uuid
        status:
          type: string
          enum: [passed, failed]
        score:
          type: integer
        canPublishListing:
          type: boolean
...